+++ tests/testthat/test-interfaces.R
@@ -5,11 +5,11 @@ test_that("NeEstimator",{
   ne_path <- "C://usr/bin/Ne2-1.exe"
   skip_if(!file.exists(ne_path))
   
-  ne <- calc_ne(stickSNPs[pop = "ASP"], NeEstimator_path = ne_path, chr = "chr", facets = "pop")
+  ne <- calc_ne(stickSNPs[pop = "ASP"], NeEstimator_path = ne_path, chr = "chr", facets = "pop", verbose = FALSE)
   ne <- get.snpR.stats(ne, "pop", "ne")
-  unlink("NeEstimator", recursive = T)
   
-  expect_snapshot_value(ne, style = "serialize") # not internally calced, just a check for proper prep and parsing
+  
+  expect_equivalent(ne$pop[1,-c(1:2), drop = TRUE], c(83.5, 24.4, Inf, 18.3, Inf)) # not internally calced, just a check for proper prep and parsing
 })
 
 
@@ -22,7 +22,7 @@ test_that("colony",{
   col_path <- "C://usr/bin/Colony/colony2s.exe"
   skip_if(!file.exists(col_path))
   
-  col <- run_colony(stickSNPs[pop = "ASP"], colony_path = col_path, run_length = 1, method = "PLS", cleanup = TRUE, verbose = FALSE)
+  .make_it_quiet(col <- run_colony(stickSNPs[pop = "ASP"], colony_path = col_path, run_length = 1, method = "PLS", cleanup = TRUE, verbose = FALSE))
   
   expect_identical(colnames(col$clusters), c("ClusterIndex", "ClusterProbability", "OffspringID", "FatherID", "MotherID"))
   snap_check <- col$dyads[col$dyads$Probability > .05, -3]